Move user model into an importable module
